UI fixes
ticket (all, new, update), project (all, new, update), login

+++ bugtracker/src/main/java/bugtracker/user/UserRepository.java
@@ -14,7 +14,7 @@ public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
     UserEntity findUserEntityByName(String name);
 
-    List<UserEntity> findByType(Long type);
+    List<UserEntity> findByType(Integer type);
 
     UserEntity findByName(String name);
